TFG_FirstTry1/requirements.txt
torch>=2.2
torchvision>=0.17
pandas>=2.0
numpy>=1.26
pillow>=10.0
tqdm>=4.66
pyyaml>=6.0
